workoutRecognition/requirements.txt
blinker==1.9.0
boto3==1.40.55
botocore==1.40.55
certifi==2025.10.5
click==8.3.0
Flask==3.1.2
flask-cors==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
MarkupSafe==3.0.3
numpy==2.2.6
opencv-python==4.12.0.88
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
s3transfer==0.14.0
six==1.17.0
urllib3==2.5.0
Werkzeug==3.1.3
